backend/requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
pymysql>=1.1.0
PyJWT>=2.8.0
flask-cors>=4.0.0
cos-python-sdk-v5>=1.9.0
